+++ CHANGES.rst
@@ -2,8 +2,8 @@ Changelog
 =========
 
 
-1.0a44 (unreleased)
--------------------
+1.0 (2022-02-22)
+----------------
 
 - Add description in sendinblue section
   [boulch]